+++ src/main/java/javax/util/streamex/MoreCollectors.java
@@ -21,7 +21,10 @@ import java.util.Comparator;
 import java.util.Deque;
 import java.util.EnumMap;
 import java.util.EnumSet;
+import java.util.LinkedHashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.Map.Entry;
 import java.util.Optional;
 import java.util.OptionalLong;
 import java.util.PriorityQueue;
@@ -31,6 +34,7 @@ import java.util.function.BiFunction;
 import java.util.function.BinaryOperator;
 import java.util.function.Function;
 import java.util.function.IntFunction;
+import java.util.function.Predicate;
 import java.util.function.Supplier;
 import java.util.stream.Collector;
 import java.util.stream.Collector.Characteristics;
@@ -94,6 +98,29 @@ public final class MoreCollectors {
     }
 
     /**
+     * Returns a {@code Collector} which produces a boolean array containing the
+     * results of applying the given predicate to the input elements, in
+     * encounter order.
+     * 
+     * @param <T>
+     *            the type of the input elements
+     * @param predicate
+     *            a non-interfering, stateless predicate to apply to each input
+     *            element. The result values of this predicate are collected to
+     *            the resulting boolean array.
+     * @return a {@code Collector} which collects the results of the predicate
+     *         function to the boolean array, in encounter order.
+     * @since 0.3.8
+     */
+    public static <T> Collector<T, ?, boolean[]> toBooleanArray(Predicate<T> predicate) {
+        return PartialCollector.booleanArray().asRef((box, t) -> {
+            if (predicate.test(t))
+                box.a.set(box.b);
+            box.b = StrictMath.addExact(box.b, 1);
+        });
+    }
+
+    /**
      * Returns a {@code Collector} which counts a number of distinct values the
      * mapper function returns for the stream elements.
      * 
@@ -103,15 +130,45 @@ public final class MoreCollectors {
      * 
      * @param <T>
      *            the type of the input elements
-     * @param <U>
-     *            the type of objects the mapper function produces
      * @param mapper
      *            a function which classifies input elements.
      * @return a collector which counts a number of distinct classes the mapper
      *         function returns for the stream elements.
      */
-    public static <T, U> Collector<T, ?, Integer> distinctCount(Function<? super T, U> mapper) {
+    public static <T> Collector<T, ?, Integer> distinctCount(Function<? super T, ?> mapper) {
         return Collectors.collectingAndThen(Collectors.mapping(mapper, Collectors.toSet()), Set::size);
+    }
+
+    /**
+     * Returns a {@code Collector} which collects into the {@link List} the
+     * input elements for which given mapper function returns distinct results.
+     *
+     * For ordered source the order of collected elements is preserved. If the
+     * same result is returned by mapper function for several elements, only the
+     * first element is included into the resulting list.
+     * 
+     * There are no guarantees on the type, mutability, serializability, or
+     * thread-safety of the {@code List} returned.
+     * 
+     * The operation performed by the returned collector is equivalent to
+     * {@code stream.distinct(mapper).toList()}. This collector is mostly useful
+     * as a downstream collector.
+     * 
+     * @param <T>
+     *            the type of the input elements
+     * @param mapper
+     *            a function which classifies input elements.
+     * @return a collector which collects distinct elements to the {@code List}.
+     * @since 0.3.8
+     */
+    public static <T> Collector<T, ?, List<T>> distinctBy(Function<? super T, ?> mapper) {
+        return Collector.<T, Map<Object, T>, List<T>> of(LinkedHashMap::new,
+            (map, t) -> map.putIfAbsent(mapper.apply(t), t), (m1, m2) -> {
+                for(Entry<Object, T> e : m2.entrySet()) {
+                    m1.putIfAbsent(e.getKey(), e.getValue());
+                }
+                return m1;
+            }, map -> new ArrayList<>(map.values()));
     }
 
     /**
@@ -126,7 +183,7 @@ public final class MoreCollectors {
      * @see Collectors#counting()
      */
     public static <T> Collector<T, ?, Integer> countingInt() {
-        return Collectors.reducing(0, t -> 1, Integer::sum);
+        return PartialCollector.intSum().asRef((acc, t) -> acc[0]++);
     }
 
     /**
@@ -498,9 +555,12 @@ public final class MoreCollectors {
         if (n <= 0)
             return empty();
         BiConsumer<PriorityQueue<T>, T> accumulator = (queue, t) -> {
-            queue.add(t);
-            if (queue.size() > n)
+            if (queue.size() < n)
+                queue.add(t);
+            else if (comparator.compare(queue.peek(), t) < 0) {
                 queue.poll();
+                queue.add(t);
+            }
         };
         return Collector.of(() -> new PriorityQueue<>(comparator), accumulator, (q1, q2) -> {
             for (T t : q2) {